+++ android/src/main/java/com/github/drydart/flutter_wasm/FlutterWasmPlugin.java
@@ -1,25 +1,50 @@
+/* This is free and unencumbered software released into the public domain. */
+
 package com.github.drydart.flutter_wasm;
+
+import com.github.drydart.flutter_wasm_vm.Flutter_wasm_vm;
+import com.github.drydart.flutter_wasm_vm.State;
 
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 
+import java.io.IOException;
+
 /** FlutterWasmPlugin */
-public class FlutterWasmPlugin implements MethodCallHandler {
+public class FlutterWasmPlugin extends FlutterMethodCallHandler {
+  private static final String TAG = "FlutterWasmPlugin";
+  public static final String CHANNEL = "flutter_wasm";
+
+  FlutterWasmPlugin(final Registrar registrar) {
+    super(registrar);
+  }
+
   /** Plugin registration. */
-  public static void registerWith(Registrar registrar) {
-    final MethodChannel channel = new MethodChannel(registrar.messenger(), "flutter_wasm");
-    channel.setMethodCallHandler(new FlutterWasmPlugin());
+  public static void registerWith(final Registrar registrar) {
+    assert(registrar != null);
+
+    (new MethodChannel(registrar.messenger(), CHANNEL))
+      .setMethodCallHandler(new FlutterWasmPlugin(registrar));
   }
 
   @Override
-  public void onMethodCall(MethodCall call, Result result) {
-    if (call.method.equals("getPlatformVersion")) {
-      result.success("Android " + android.os.Build.VERSION.RELEASE);
-    } else {
-      result.notImplemented();
+  public void onMethodCall(final MethodCall call, final Result result) {
+    assert(call != null);
+    assert(result != null);
+
+    assert(call.method != null);
+    switch (call.method) {
+
+      case "getVersion": {
+        result.success(Flutter_wasm_vm.version());
+        break;
+      }
+
+      default: {
+        result.notImplemented();
+      }
     }
   }
 }