+++ flutter_wasm_vm.go
@@ -1,0 +1,12 @@
+/* This is free and unencumbered software released into the public domain. */
+
+// Package flutter_wasm_vm
+package flutter_wasm_vm
+
+// State
+type State struct {}
+
+// Version
+func Version() string {
+	return "0.0.0"
+}