+++ CHANGELOG.md
@@ -8,5 +8,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2019-04-24
 ### Added
+- `Wasm.version` static getter
 
 [0.1.0]:  https://github.com/drydart/flutter_wasm/compare/0.0.0...0.1.0